+++ lib/deltachat.ts
@@ -259,16 +259,15 @@ export class DeltaChat extends EventEmitter {
     return binding.dcn_create_contact(this.dcn_context, name, addr)
   }
 
-  /** @returns chatId */
-  createUnverifiedGroupChat(chatName: string): number {
-    debug(`createUnverifiedGroupChat ${chatName}`)
-    return binding.dcn_create_group_chat(this.dcn_context, 0, chatName)
-  }
-
-  /** @returns chatId */
-  createVerifiedGroupChat(chatName: string): number {
-    debug(`createVerifiedGroupChat ${chatName}`)
-    return binding.dcn_create_group_chat(this.dcn_context, 1, chatName)
+  /**
+   * 
+   * @param chatName The name of the chat that should be created
+   * @param is_protected Whether the chat should be protected at creation time
+   * @returns chatId
+   */
+  createGroupChat(chatName: string, is_protected:boolean = false): number {
+    debug(`createGroupChat ${chatName} [protected:${is_protected}]`)
+    return binding.dcn_create_group_chat(this.dcn_context, is_protected?1:0, chatName)
   }
 
   deleteChat(chatId: number) {
@@ -739,6 +738,17 @@ export class DeltaChat extends EventEmitter {
     return Boolean(
       binding.dcn_set_chat_name(this.dcn_context, Number(chatId), name)
     )
+  }
+
+  /**
+   * 
+   * @param chatId 
+   * @param protect 
+   * @returns success boolean
+   */
+  setChatProtection(chatId:number, protect:boolean) {
+    debug(`setChatProtection ${chatId} ${protect}`)
+    return Boolean(binding.dcn_set_chat_protection(this.dcn_context, Number(chatId), protect?1:0))
   }
 
   getChatEphemeralTimer(chatId: number): number {