+++ src/module.c
@@ -550,11 +550,11 @@ NAPI_METHOD(dcn_create_contact) {
 NAPI_METHOD(dcn_create_group_chat) {
   NAPI_ARGV(3);
   NAPI_DCN_CONTEXT();
-  NAPI_ARGV_INT32(verified, 1);
+  NAPI_ARGV_INT32(protect, 1);
   NAPI_ARGV_UTF8_MALLOC(chat_name, 2);
 
   //TRACE("calling..");
-  uint32_t chat_id = dc_create_group_chat(dcn_context->dc_context, verified, chat_name);
+  uint32_t chat_id = dc_create_group_chat(dcn_context->dc_context, protect, chat_name);
   //TRACE("result %d", chat_id);
 
   free(chat_name);
@@ -1291,6 +1291,18 @@ NAPI_METHOD(dcn_set_chat_name) {
   NAPI_RETURN_INT32(result);
 }
 
+NAPI_METHOD(dcn_set_chat_protection) {
+  NAPI_ARGV(3);
+  NAPI_DCN_CONTEXT();
+  NAPI_ARGV_UINT32(chat_id, 1);
+  NAPI_ARGV_INT32(protect, 1);
+
+  int result = dc_set_chat_protection(dcn_context->dc_context,
+                                      chat_id,
+                                      protect);
+  NAPI_RETURN_INT32(result);
+}
+
 NAPI_METHOD(dcn_get_chat_ephemeral_timer) {
   NAPI_ARGV(2);
   NAPI_DCN_CONTEXT();
@@ -1532,15 +1544,15 @@ NAPI_METHOD(dcn_chat_is_unpromoted) {
   NAPI_RETURN_INT32(is_unpromoted);
 }
 
-NAPI_METHOD(dcn_chat_is_verified) {
+NAPI_METHOD(dcn_chat_is_protected) {
   NAPI_ARGV(1);
   NAPI_DC_CHAT();
 
   //TRACE("calling..");
-  int is_verified = dc_chat_is_verified(dc_chat);
-  //TRACE("result %d", is_verified);
+  int is_protected = dc_chat_is_protected(dc_chat);
+  //TRACE("result %d", is_protected);
 
-  NAPI_RETURN_INT32(is_verified);
+  NAPI_RETURN_INT32(is_protected);
 }
 
 NAPI_METHOD(dcn_chat_is_device_talk) {
@@ -2595,6 +2607,7 @@ NAPI_INIT() {
   NAPI_EXPORT_FUNCTION(dcn_send_msg);
   NAPI_EXPORT_FUNCTION(dcn_send_videochat_invitation);
   NAPI_EXPORT_FUNCTION(dcn_set_chat_name);
+  NAPI_EXPORT_FUNCTION(dcn_set_chat_protection);
   NAPI_EXPORT_FUNCTION(dcn_get_chat_ephemeral_timer);
   NAPI_EXPORT_FUNCTION(dcn_set_chat_ephemeral_timer);
   NAPI_EXPORT_FUNCTION(dcn_set_chat_profile_image);
@@ -2620,7 +2633,7 @@ NAPI_INIT() {
   NAPI_EXPORT_FUNCTION(dcn_chat_get_type);
   NAPI_EXPORT_FUNCTION(dcn_chat_is_self_talk);
   NAPI_EXPORT_FUNCTION(dcn_chat_is_unpromoted);
-  NAPI_EXPORT_FUNCTION(dcn_chat_is_verified);
+  NAPI_EXPORT_FUNCTION(dcn_chat_is_protected);
   NAPI_EXPORT_FUNCTION(dcn_chat_is_device_talk);
   NAPI_EXPORT_FUNCTION(dcn_chat_is_muted);